+++ src/handler.rs
@@ -2,8 +2,10 @@ mod core;
 mod endpoint;
 mod filter;
 mod filter_map;
+mod map;
 
 pub use self::core::*;
 pub use endpoint::*;
 pub use filter::*;
 pub use filter_map::*;
+pub use map::*;