+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
  - `DependencyMap::insert_container` ([PR #7](https://github.com/p0lunin/dptree/pull/7)).
+ - `map` and `map_async` ([PR #8](https://github.com/p0lunin/dptree/pull/8)).
 
 ## 0.1.1 - 2022-03-21