+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## unreleased
 
+## 0.1.2 - 2022-04-04
+
 ### Added
 
  - `DependencyMap::insert_container` ([PR #7](https://github.com/p0lunin/dptree/pull/7)).